pyproject.toml
[tool.poetry]
name = "todocli"
version = "0.1.0"
description = ""
authors = ["sarrrrryy <sarrrrry.sarrrrry@gmail.com>"]
readme = "README.md"
packages = [
    {include = 'src'},
]

[tool.poetry.dependencies]
python = "^3.10"
pytodoist = "^2.1.4"
python-dotenv = "^0.21.0"
click = "^8.1.3"
yaspin = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
